feat: refactor and clean up
